services/perception/requirements.txt
ultralytics>=8.3.0
opencv-python-headless==4.9.0.80
paho-mqtt==2.1.0
python-dotenv==1.0.0
pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0.0
